content: refactored content manager
- separated portions that don't require locking into separate struct
  to make it easier to reason about state
- moved iteration-related content to separate file
