+++ repo/content/content_manager.go
@@ -4,18 +4,12 @@ package content
 import (
 	"bytes"
 	"context"
-	"crypto/aes"
 	cryptorand "crypto/rand"
 	"encoding/hex"
 	"fmt"
-	"io"
-	"math"
-	"math/rand"
 	"os"
 	"reflect"
-	"strings"
 	"sync"
-	"sync/atomic"
 	"time"
 
 	"github.com/pkg/errors"
@@ -71,40 +65,17 @@ type IndexBlobInfo struct {
 
 // Manager builds content-addressable storage with encryption, deduplication and packaging on top of BLOB store.
 type Manager struct {
-	Format         FormattingOptions
-	CachingOptions CachingOptions
+	mu     sync.Mutex
+	locked bool
 
-	stats         Stats
-	contentCache  *contentCache
-	metadataCache *contentCache
-	listCache     *listCache
-	st            blob.Storage
-
-	mu                      sync.Mutex
-	locked                  bool
-	checkInvariantsOnUnlock bool
-
-	pendingPacks      map[blob.ID]*pendingPackInfo
-	packIndexBuilder  packIndexBuilder // contents that are in index currently being built (current pack and all packs saved but not committed)
-	committedContents *committedContentIndex
+	pendingPacks     map[blob.ID]*pendingPackInfo
+	packIndexBuilder packIndexBuilder // contents that are in index currently being built (current pack and all packs saved but not committed)
 
 	disableIndexFlushCount int
 	flushPackIndexesAfter  time.Time // time when those indexes should be flushed
+	closed                 chan struct{}
 
-	closed chan struct{}
-
-	writeFormatVersion int32 // format version to write
-
-	maxPackSize int
-	hasher      HashFunc
-	encryptor   Encryptor
-
-	minPreambleLength int
-	maxPreambleLength int
-	paddingUnit       int
-	timeNow           func() time.Time
-
-	repositoryFormatBytes []byte
+	lockFreeManager
 }
 
 type pendingPackInfo struct {
@@ -139,7 +110,7 @@ func (bm *Manager) DeleteContent(contentID ID) error {
 			bi2 := *bi
 			bi2.Deleted = true
 			bi2.TimestampSeconds = bm.timeNow().Unix()
-			bm.setPendingContent(bm.getOrCreatePendingPackInfoLocked(bm.packPrefixForContentID(contentID)), bi2)
+			bm.setPendingContent(bm.getOrCreatePendingPackInfoLocked(packPrefixForContentID(contentID)), bi2)
 		}
 		return nil
 	}
@@ -159,7 +130,7 @@ func (bm *Manager) DeleteContent(contentID ID) error {
 	bi2 := bi
 	bi2.Deleted = true
 	bi2.TimestampSeconds = bm.timeNow().Unix()
-	bm.setPendingContent(bm.getOrCreatePendingPackInfoLocked(bm.packPrefixForContentID(contentID)), bi2)
+	bm.setPendingContent(bm.getOrCreatePendingPackInfoLocked(packPrefixForContentID(contentID)), bi2)
 	return nil
 }
 
@@ -173,7 +144,7 @@ func (bm *Manager) setPendingContent(pp *pendingPackInfo, i Info) {
 func (bm *Manager) addToPackLocked(ctx context.Context, contentID ID, data []byte, isDeleted bool) error {
 	bm.assertLocked()
 
-	prefix := bm.packPrefixForContentID(contentID)
+	prefix := packPrefixForContentID(contentID)
 	pp := bm.getOrCreatePendingPackInfoLocked(prefix)
 
 	data = cloneBytes(data)
@@ -324,10 +295,6 @@ func (bm *Manager) flushPackIndexesLocked(ctx context.Context) error {
 	return nil
 }
 
-func (bm *Manager) writePackIndexesNew(ctx context.Context, data []byte) (blob.ID, error) {
-	return bm.encryptAndWriteContentNotLocked(ctx, data, newIndexBlobPrefix)
-}
-
 func (bm *Manager) finishAllPacksLocked(ctx context.Context) error {
 	for prefix, pp := range bm.pendingPacks {
 		if len(pp.currentPackItems) == 0 {
@@ -373,352 +340,11 @@ func (bm *Manager) finishPackLocked(ctx context.Context, prefix blob.ID, pp *pen
 	return nil
 }
 
-func (bm *Manager) preparePackDataContent(ctx context.Context, pp *pendingPackInfo, packFile blob.ID) ([]byte, packIndexBuilder, error) {
-	formatLog.Debugf("preparing content data with %v items", len(pp.currentPackItems))
-
-	contentData, err := appendRandomBytes(append([]byte(nil), bm.repositoryFormatBytes...), rand.Intn(bm.maxPreambleLength-bm.minPreambleLength+1)+bm.minPreambleLength)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "unable to prepare content preamble")
-	}
-
-	packFileIndex := packIndexBuilder{}
-	for contentID, info := range pp.currentPackItems {
-		if info.Payload == nil {
-			continue
-		}
-
-		var encrypted []byte
-		encrypted, err = bm.maybeEncryptContentDataForPacking(info.Payload, info.ID)
-		if err != nil {
-			return nil, nil, errors.Wrapf(err, "unable to encrypt %q", contentID)
-		}
-
-		formatLog.Debugf("adding %v length=%v deleted=%v", contentID, len(info.Payload), info.Deleted)
-
-		packFileIndex.Add(Info{
-			ID:               contentID,
-			Deleted:          info.Deleted,
-			FormatVersion:    byte(bm.writeFormatVersion),
-			PackBlobID:       packFile,
-			PackOffset:       uint32(len(contentData)),
-			Length:           uint32(len(encrypted)),
-			TimestampSeconds: info.TimestampSeconds,
-		})
-
-		if contentID.HasPrefix() {
-			bm.metadataCache.put(ctx, cacheKey(contentID), cloneBytes(encrypted))
-		}
-
-		contentData = append(contentData, encrypted...)
-	}
-
-	if len(packFileIndex) == 0 {
-		return nil, nil, nil
-	}
-
-	if bm.paddingUnit > 0 {
-		if missing := bm.paddingUnit - (len(contentData) % bm.paddingUnit); missing > 0 {
-			contentData, err = appendRandomBytes(contentData, missing)
-			if err != nil {
-				return nil, nil, errors.Wrap(err, "unable to prepare content postamble")
-			}
-		}
-	}
-
-	origContentLength := len(contentData)
-	contentData, err = bm.appendPackFileIndexRecoveryData(contentData, packFileIndex)
-
-	formatLog.Debugf("finished content %v bytes (%v bytes index)", len(contentData), len(contentData)-origContentLength)
-	return contentData, packFileIndex, err
-}
-
-func (bm *Manager) maybeEncryptContentDataForPacking(data []byte, contentID ID) ([]byte, error) {
-	iv, err := getPackedContentIV(contentID)
-	if err != nil {
-		return nil, errors.Wrapf(err, "unable to get packed content IV for %q", contentID)
-	}
-
-	return bm.encryptor.Encrypt(data, iv)
-}
-
-func appendRandomBytes(b []byte, count int) ([]byte, error) {
-	rnd := make([]byte, count)
-	if _, err := io.ReadFull(cryptorand.Reader, rnd); err != nil {
-		return nil, err
-	}
-
-	return append(b, rnd...), nil
-}
-
-// IndexBlobs returns the list of active index blobs.
-func (bm *Manager) IndexBlobs(ctx context.Context) ([]IndexBlobInfo, error) {
-	return bm.listCache.listIndexBlobs(ctx)
-}
-
-func (bm *Manager) loadPackIndexesUnlocked(ctx context.Context) ([]IndexBlobInfo, bool, error) {
-	nextSleepTime := 100 * time.Millisecond
-
-	for i := 0; i < indexLoadAttempts; i++ {
-		if err := ctx.Err(); err != nil {
-			return nil, false, err
-		}
-
-		if i > 0 {
-			bm.listCache.deleteListCache()
-			log.Debugf("encountered NOT_FOUND when loading, sleeping %v before retrying #%v", nextSleepTime, i)
-			time.Sleep(nextSleepTime)
-			nextSleepTime *= 2
-		}
-
-		contents, err := bm.listCache.listIndexBlobs(ctx)
-		if err != nil {
-			return nil, false, err
-		}
-
-		err = bm.tryLoadPackIndexBlobsUnlocked(ctx, contents)
-		if err == nil {
-			var contentIDs []blob.ID
-			for _, b := range contents {
-				contentIDs = append(contentIDs, b.BlobID)
-			}
-			var updated bool
-			updated, err = bm.committedContents.use(contentIDs)
-			if err != nil {
-				return nil, false, err
-			}
-			return contents, updated, nil
-		}
-		if err != blob.ErrBlobNotFound {
-			return nil, false, err
-		}
-	}
-
-	return nil, false, errors.Errorf("unable to load pack indexes despite %v retries", indexLoadAttempts)
-}
-
-func (bm *Manager) tryLoadPackIndexBlobsUnlocked(ctx context.Context, contents []IndexBlobInfo) error {
-	ch, unprocessedIndexesSize, err := bm.unprocessedIndexBlobsUnlocked(contents)
-	if err != nil {
-		return err
-	}
-	if len(ch) == 0 {
-		return nil
-	}
-
-	log.Infof("downloading %v new index blobs (%v bytes)...", len(ch), unprocessedIndexesSize)
-	var wg sync.WaitGroup
-
-	errch := make(chan error, parallelFetches)
-
-	for i := 0; i < parallelFetches; i++ {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-
-			for indexBlobID := range ch {
-				data, err := bm.getIndexBlobInternal(ctx, indexBlobID)
-				if err != nil {
-					errch <- err
-					return
-				}
-
-				if err := bm.committedContents.addContent(indexBlobID, data, false); err != nil {
-					errch <- errors.Wrap(err, "unable to add to committed content cache")
-					return
-				}
-			}
-		}()
-	}
-
-	wg.Wait()
-	close(errch)
-
-	// Propagate async errors, if any.
-	for err := range errch {
-		return err
-	}
-	log.Infof("Index contents downloaded.")
-
-	return nil
-}
-
-// unprocessedIndexBlobsUnlocked returns a closed channel filled with content IDs that are not in committedContents cache.
-func (bm *Manager) unprocessedIndexBlobsUnlocked(contents []IndexBlobInfo) (resultCh <-chan blob.ID, totalSize int64, err error) {
-	ch := make(chan blob.ID, len(contents))
-	for _, c := range contents {
-		has, err := bm.committedContents.cache.hasIndexBlobID(c.BlobID)
-		if err != nil {
-			return nil, 0, err
-		}
-		if has {
-			log.Debugf("index blob %q already in cache, skipping", c.BlobID)
-			continue
-		}
-		ch <- c.BlobID
-		totalSize += c.Length
-	}
-	close(ch)
-	return ch, totalSize, nil
-}
-
 // Close closes the content manager.
 func (bm *Manager) Close() {
 	bm.contentCache.close()
 	bm.metadataCache.close()
 	close(bm.closed)
-}
-
-type IterateOptions struct {
-	Prefix         ID
-	IncludeDeleted bool
-	Parallel       int
-}
-
-type IterateCallback func(Info) error
-type cancelIterateFunc func() error
-
-func maybeParallelExecutor(parallel int, originalCallback IterateCallback) (IterateCallback, cancelIterateFunc) {
-	if parallel <= 1 {
-		return originalCallback, func() error { return nil }
-	}
-
-	workch := make(chan Info, parallel)
-	workererrch := make(chan error, 1)
-	var wg sync.WaitGroup
-	var once sync.Once
-
-	lastWorkerError := func() error {
-		select {
-		case err := <-workererrch:
-			return err
-		default:
-			return nil
-		}
-	}
-
-	cleanup := func() error {
-		once.Do(func() {
-			close(workch)
-			wg.Wait()
-		})
-		return lastWorkerError()
-	}
-
-	callback := func(i Info) error {
-		workch <- i
-		return lastWorkerError()
-	}
-
-	// start N workers, each fetching from the shared channel and invoking the provided callback.
-	// cleanup() must be called to for worker completion
-	for i := 0; i < parallel; i++ {
-		wg.Add(1)
-
-		go func() {
-			defer wg.Done()
-			for i := range workch {
-				if err := originalCallback(i); err != nil {
-					select {
-					case workererrch <- err:
-					default:
-					}
-				}
-			}
-		}()
-	}
-
-	return callback, cleanup
-}
-
-// IterateContents invokes the provided callback for each content starting with a specified prefix
-// and possibly including deleted items.
-func (bm *Manager) IterateContents(opts IterateOptions, callback IterateCallback) error {
-	bm.lock()
-	pibClone := bm.packIndexBuilder.clone()
-	bm.unlock()
-
-	callback, cleanup := maybeParallelExecutor(opts.Parallel, callback)
-	defer cleanup() //nolint:errcheck
-
-	invokeCallback := func(i Info) error {
-		if !opts.IncludeDeleted {
-			if ci, ok := pibClone[i.ID]; ok {
-				if ci.Deleted {
-					return nil
-				}
-			} else if i.Deleted {
-				return nil
-			}
-		}
-
-		if !strings.HasPrefix(string(i.ID), string(opts.Prefix)) {
-			return nil
-		}
-		return callback(i)
-	}
-
-	if len(pibClone) == 0 && opts.IncludeDeleted && opts.Prefix == "" && opts.Parallel <= 1 {
-		// fast path, invoke callback directly
-		invokeCallback = callback
-	}
-
-	for _, bi := range pibClone {
-		_ = invokeCallback(*bi)
-	}
-
-	if err := bm.committedContents.listContents(opts.Prefix, invokeCallback); err != nil {
-		return err
-	}
-
-	return cleanup()
-}
-
-type IteratePackOptions struct {
-	IncludePacksWithOnlyDeletedContent bool
-	IncludeContentInfos                bool
-}
-
-type PackInfo struct {
-	PackID       blob.ID
-	ContentCount int
-	TotalSize    int64
-	ContentInfos []Info
-}
-
-type IteratePacksCallback func(PackInfo) error
-
-// IteratePacks invokes the provided callback for all pack blobs.
-func (bm *Manager) IteratePacks(options IteratePackOptions, callback IteratePacksCallback) error {
-	packUsage := map[blob.ID]*PackInfo{}
-
-	if err := bm.IterateContents(
-		IterateOptions{
-			IncludeDeleted: options.IncludePacksWithOnlyDeletedContent,
-		},
-		func(ci Info) error {
-			pi := packUsage[ci.PackBlobID]
-			if pi == nil {
-				pi = &PackInfo{}
-				packUsage[ci.PackBlobID] = pi
-			}
-			pi.PackID = ci.PackBlobID
-			pi.ContentCount++
-			pi.TotalSize += int64(ci.Length)
-			if options.IncludeContentInfos {
-				pi.ContentInfos = append(pi.ContentInfos, ci)
-			}
-			return nil
-		}); err != nil {
-		return errors.Wrap(err, "error iterating contents")
-	}
-
-	for _, v := range packUsage {
-		if err := callback(*v); err != nil {
-			return err
-		}
-	}
-
-	return nil
 }
 
 // Flush completes writing any pending packs and writes pack indexes to the underlyign storage.
@@ -755,7 +381,7 @@ func (bm *Manager) RewriteContent(ctx context.Context, contentID ID) error {
 	return bm.addToPackLocked(ctx, contentID, data, bi.Deleted)
 }
 
-func (bm *Manager) packPrefixForContentID(contentID ID) blob.ID {
+func packPrefixForContentID(contentID ID) blob.ID {
 	if contentID.HasPrefix() {
 		return PackBlobIDPrefixSpecial
 	}
@@ -792,59 +418,6 @@ func (bm *Manager) WriteContent(ctx context.Context, data []byte, prefix ID) (ID
 	defer bm.unlock()
 	err := bm.addToPackLocked(ctx, contentID, data, false)
 	return contentID, err
-}
-
-func validatePrefix(prefix ID) error {
-	switch len(prefix) {
-	case 0:
-		return nil
-	case 1:
-		if prefix[0] >= 'g' && prefix[0] <= 'z' {
-			return nil
-		}
-	}
-
-	return errors.Errorf("invalid prefix, must be a empty or single letter between 'g' and 'z'")
-}
-
-func (bm *Manager) writePackFileNotLocked(ctx context.Context, packFile blob.ID, data []byte) error {
-	atomic.AddInt32(&bm.stats.WrittenContents, 1)
-	atomic.AddInt64(&bm.stats.WrittenBytes, int64(len(data)))
-	bm.listCache.deleteListCache()
-	return bm.st.PutBlob(ctx, packFile, data)
-}
-
-func (bm *Manager) encryptAndWriteContentNotLocked(ctx context.Context, data []byte, prefix blob.ID) (blob.ID, error) {
-	hash := bm.hashData(data)
-	blobID := prefix + blob.ID(hex.EncodeToString(hash))
-
-	// Encrypt the content in-place.
-	atomic.AddInt64(&bm.stats.EncryptedBytes, int64(len(data)))
-	data2, err := bm.encryptor.Encrypt(data, hash)
-	if err != nil {
-		return "", err
-	}
-
-	atomic.AddInt32(&bm.stats.WrittenContents, 1)
-	atomic.AddInt64(&bm.stats.WrittenBytes, int64(len(data2)))
-	bm.listCache.deleteListCache()
-	if err := bm.st.PutBlob(ctx, blobID, data2); err != nil {
-		return "", err
-	}
-
-	return blobID, nil
-}
-
-func (bm *Manager) hashData(data []byte) []byte {
-	// Hash the content and compute encryption key.
-	contentID := bm.hasher(data)
-	atomic.AddInt32(&bm.stats.HashedContents, 1)
-	atomic.AddInt64(&bm.stats.HashedBytes, int64(len(data)))
-	return contentID
-}
-
-func cloneBytes(b []byte) []byte {
-	return append([]byte{}, b...)
 }
 
 // GetContent gets the contents of a given content. If the content is not found returns ErrContentNotFound.
@@ -901,186 +474,6 @@ func (bm *Manager) ContentInfo(ctx context.Context, contentID ID) (Info, error) 
 	return bi, err
 }
 
-// IterateContentInShortPacks invokes the provided callback for all contents that are stored in
-// packs shorter than the given threshold.
-func (bm *Manager) IterateContentInShortPacks(threshold int64, callback IterateCallback) error {
-	if threshold <= 0 {
-		threshold = int64(bm.maxPackSize) * 8 / 10
-	}
-
-	return bm.IteratePacks(
-		IteratePackOptions{
-			IncludePacksWithOnlyDeletedContent: true,
-			IncludeContentInfos:                true,
-		},
-		func(pi PackInfo) error {
-			if pi.TotalSize >= threshold {
-				return nil
-			}
-
-			for _, ci := range pi.ContentInfos {
-				if err := callback(ci); err != nil {
-					return err
-				}
-			}
-			return nil
-		},
-	)
-}
-
-// FindUnreferencedBlobs returns the list of unreferenced storage blobs.
-func (bm *Manager) IterateUnreferencedBlobs(ctx context.Context, parallellism int, callback func(blob.Metadata) error) error {
-	usedPacks := map[blob.ID]bool{}
-
-	log.Infof("determining blobs in use")
-	// find packs in use
-	if err := bm.IteratePacks(
-		IteratePackOptions{
-			IncludePacksWithOnlyDeletedContent: true,
-		},
-		func(pi PackInfo) error {
-			if pi.ContentCount > 0 {
-				usedPacks[pi.PackID] = true
-			}
-			return nil
-		}); err != nil {
-		return errors.Wrap(err, "error iterating packs")
-	}
-	log.Infof("found %v pack blobs in use", len(usedPacks))
-
-	unusedCount := 0
-	var prefixes []blob.ID
-
-	if parallellism <= len(PackBlobIDPrefixes) {
-		prefixes = append(prefixes, PackBlobIDPrefixes...)
-	} else {
-		// iterate {p,q}[0-9,a-f]
-		for _, prefix := range PackBlobIDPrefixes {
-			for hexDigit := 0; hexDigit < 16; hexDigit++ {
-				prefixes = append(prefixes, blob.ID(fmt.Sprintf("%v%x", prefix, hexDigit)))
-			}
-		}
-	}
-	if err := blob.IterateAllPrefixesInParallel(ctx, parallellism, bm.st, prefixes,
-		func(bm blob.Metadata) error {
-			if usedPacks[bm.BlobID] {
-				return nil
-			}
-
-			unusedCount++
-			return callback(bm)
-		}); err != nil {
-		return errors.Wrap(err, "error iterating blobs")
-	}
-	log.Infof("found %v pack blobs not in use", unusedCount)
-
-	return nil
-}
-
-func (bm *Manager) getCacheForContentID(id ID) *contentCache {
-	if id.HasPrefix() {
-		return bm.metadataCache
-	}
-
-	return bm.contentCache
-}
-
-func (bm *Manager) getContentDataUnlocked(ctx context.Context, bi *Info) ([]byte, error) {
-	if bi.Payload != nil {
-		return cloneBytes(bi.Payload), nil
-	}
-
-	payload, err := bm.getCacheForContentID(bi.ID).getContent(ctx, cacheKey(bi.ID), bi.PackBlobID, int64(bi.PackOffset), int64(bi.Length))
-	if err != nil {
-		return nil, err
-	}
-
-	atomic.AddInt32(&bm.stats.ReadContents, 1)
-	atomic.AddInt64(&bm.stats.ReadBytes, int64(len(payload)))
-
-	iv, err := getPackedContentIV(bi.ID)
-	if err != nil {
-		return nil, err
-	}
-
-	decrypted, err := bm.decryptAndVerify(payload, iv)
-	if err != nil {
-		return nil, errors.Wrapf(err, "invalid checksum at %v offset %v length %v", bi.PackBlobID, bi.PackOffset, len(payload))
-	}
-
-	return decrypted, nil
-}
-
-func (bm *Manager) decryptAndVerify(encrypted, iv []byte) ([]byte, error) {
-	decrypted, err := bm.encryptor.Decrypt(encrypted, iv)
-	if err != nil {
-		return nil, errors.Wrap(err, "decrypt")
-	}
-
-	atomic.AddInt64(&bm.stats.DecryptedBytes, int64(len(decrypted)))
-
-	if bm.encryptor.IsAuthenticated() {
-		// already verified
-		return decrypted, nil
-	}
-
-	// Since the encryption key is a function of data, we must be able to generate exactly the same key
-	// after decrypting the content. This serves as a checksum.
-	return decrypted, bm.verifyChecksum(decrypted, iv)
-}
-
-func (bm *Manager) getIndexBlobInternal(ctx context.Context, blobID blob.ID) ([]byte, error) {
-	payload, err := bm.contentCache.getContent(ctx, cacheKey(blobID), blobID, 0, -1)
-	if err != nil {
-		return nil, err
-	}
-
-	iv, err := getIndexBlobIV(blobID)
-	if err != nil {
-		return nil, err
-	}
-
-	atomic.AddInt32(&bm.stats.ReadContents, 1)
-	atomic.AddInt64(&bm.stats.ReadBytes, int64(len(payload)))
-
-	payload, err = bm.encryptor.Decrypt(payload, iv)
-	atomic.AddInt64(&bm.stats.DecryptedBytes, int64(len(payload)))
-	if err != nil {
-		return nil, err
-	}
-
-	// Since the encryption key is a function of data, we must be able to generate exactly the same key
-	// after decrypting the content. This serves as a checksum.
-	if err := bm.verifyChecksum(payload, iv); err != nil {
-		return nil, err
-	}
-
-	return payload, nil
-}
-
-func getPackedContentIV(contentID ID) ([]byte, error) {
-	return hex.DecodeString(string(contentID[len(contentID)-(aes.BlockSize*2):]))
-}
-
-func getIndexBlobIV(s blob.ID) ([]byte, error) {
-	if p := strings.Index(string(s), "-"); p >= 0 { // nolint:gocritic
-		s = s[0:p]
-	}
-	return hex.DecodeString(string(s[len(s)-(aes.BlockSize*2):]))
-}
-
-func (bm *Manager) verifyChecksum(data, contentID []byte) error {
-	expected := bm.hasher(data)
-	expected = expected[len(expected)-aes.BlockSize:]
-	if !bytes.HasSuffix(contentID, expected) {
-		atomic.AddInt32(&bm.stats.InvalidContents, 1)
-		return errors.Errorf("invalid checksum for blob %x, expected %x", contentID, expected)
-	}
-
-	atomic.AddInt32(&bm.stats.ValidContents, 1)
-	return nil
-}
-
 func (bm *Manager) lock() {
 	bm.mu.Lock()
 	bm.locked = true
@@ -1111,32 +504,6 @@ func (bm *Manager) Refresh(ctx context.Context) (bool, error) {
 	_, updated, err := bm.loadPackIndexesUnlocked(ctx)
 	log.Debugf("Refresh completed in %v and updated=%v", time.Since(t0), updated)
 	return updated, err
-}
-
-type cachedList struct {
-	Timestamp time.Time       `json:"timestamp"`
-	Contents  []IndexBlobInfo `json:"contents"`
-}
-
-// listIndexBlobsFromStorage returns the list of index blobs in the given storage.
-// The list of contents is not guaranteed to be sorted.
-func listIndexBlobsFromStorage(ctx context.Context, st blob.Storage) ([]IndexBlobInfo, error) {
-	snapshot, err := blob.ListAllBlobsConsistent(ctx, st, newIndexBlobPrefix, math.MaxInt32)
-	if err != nil {
-		return nil, err
-	}
-
-	var results []IndexBlobInfo
-	for _, it := range snapshot {
-		ii := IndexBlobInfo{
-			BlobID:    it.BlobID,
-			Timestamp: it.Timestamp,
-			Length:    it.Length,
-		}
-		results = append(results, ii)
-	}
-
-	return results, err
 }
 
 // NewManager creates new content manager with given packing options and a formatter.
@@ -1181,28 +548,30 @@ func newManagerWithOptions(ctx context.Context, st blob.Storage, f *FormattingOp
 	contentIndex := newCommittedContentIndex(caching)
 
 	m := &Manager{
-		Format:                *f,
-		CachingOptions:        caching,
-		timeNow:               timeNow,
+		lockFreeManager: lockFreeManager{
+			Format:                  *f,
+			CachingOptions:          caching,
+			timeNow:                 timeNow,
+			maxPackSize:             f.MaxPackSize,
+			encryptor:               encryptor,
+			hasher:                  hasher,
+			minPreambleLength:       defaultMinPreambleLength,
+			maxPreambleLength:       defaultMaxPreambleLength,
+			paddingUnit:             defaultPaddingUnit,
+			contentCache:            contentCache,
+			metadataCache:           metadataCache,
+			listCache:               listCache,
+			st:                      st,
+			repositoryFormatBytes:   repositoryFormatBytes,
+			checkInvariantsOnUnlock: os.Getenv("KOPIA_VERIFY_INVARIANTS") != "",
+			writeFormatVersion:      int32(f.Version),
+			committedContents:       contentIndex,
+		},
+
 		flushPackIndexesAfter: timeNow().Add(flushPackIndexTimeout),
-		maxPackSize:           f.MaxPackSize,
-		encryptor:             encryptor,
-		hasher:                hasher,
 		pendingPacks:          map[blob.ID]*pendingPackInfo{},
 		packIndexBuilder:      make(packIndexBuilder),
-		committedContents:     contentIndex,
-		minPreambleLength:     defaultMinPreambleLength,
-		maxPreambleLength:     defaultMaxPreambleLength,
-		paddingUnit:           defaultPaddingUnit,
-		contentCache:          contentCache,
-		metadataCache:         metadataCache,
-		listCache:             listCache,
-		st:                    st,
-		repositoryFormatBytes: repositoryFormatBytes,
-
-		writeFormatVersion:      int32(f.Version),
-		closed:                  make(chan struct{}),
-		checkInvariantsOnUnlock: os.Getenv("KOPIA_VERIFY_INVARIANTS") != "",
+		closed:                make(chan struct{}),
 	}
 
 	if err := m.CompactIndexes(ctx, autoCompactionOptions); err != nil {
@@ -1210,51 +579,4 @@ func newManagerWithOptions(ctx context.Context, st blob.Storage, f *FormattingOp
 	}
 
 	return m, nil
-}
-
-func CreateHashAndEncryptor(f *FormattingOptions) (HashFunc, Encryptor, error) {
-	h, err := createHashFunc(f)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "unable to create hash")
-	}
-
-	e, err := createEncryptor(f)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "unable to create encryptor")
-	}
-
-	contentID := h(nil)
-	_, err = e.Encrypt(nil, contentID)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "invalid encryptor")
-	}
-
-	return h, e, nil
-}
-
-func createHashFunc(f *FormattingOptions) (HashFunc, error) {
-	h := hashFunctions[f.Hash]
-	if h == nil {
-		return nil, errors.Errorf("unknown hash function %v", f.Hash)
-	}
-
-	hashFunc, err := h(f)
-	if err != nil {
-		return nil, errors.Wrap(err, "unable to initialize hash")
-	}
-
-	if hashFunc == nil {
-		return nil, errors.Errorf("nil hash function returned for %v", f.Hash)
-	}
-
-	return hashFunc, nil
-}
-
-func createEncryptor(f *FormattingOptions) (Encryptor, error) {
-	e := encryptors[f.Encryption]
-	if e == nil {
-		return nil, errors.Errorf("unknown encryption algorithm: %v", f.Encryption)
-	}
-
-	return e(f)
 }
